sonic_python_inference/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sonic_python_inference"
version = "0.1.0"
description = "SONIC VR 3PT Python inference pipeline (Isaac Lab, batched)"
requires-python = "==3.11.*"
dependencies = [
    # Core numerics
    "numpy>=1.26,<2",
    "torch==2.7.0",
    "torchvision==0.22.0",

    # ONNX runtime + export helpers
    "onnx",
    "onnxruntime-gpu",
    "vector-quantize-pytorch",

    # Unpickling sonic_release/last.pt needs these exact classes
    "trl==0.28.0",
    "transformers>=4.56.2",
    "accelerate>=1.3.0",
    "hydra-core==1.3.2",

    # Isaac Lab (pulls Isaac Sim on first run; ~10 min extension fetch)
    "isaaclab[isaacsim,all]==2.3.2.post1",

    # Optional command sources
    "pyzmq",
    "msgpack",
    "msgpack-numpy",
]

[tool.setuptools.packages.find]
where = [".."]
include = ["sonic_python_inference*"]

[tool.uv]
# `uv sync` / `uv add` pin these indices. `uv pip install -e .` does NOT honor
# `[tool.uv.sources]`; see README for the explicit `--extra-index-url` invocation.

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "nvidia"
url = "https://pypi.nvidia.com"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
isaaclab = { index = "nvidia" }
isaacsim = { index = "nvidia" }
